refactor: reorganize agent-related modules and remove deprecated LLM provider
